armenian-bank-voice-agent/requirements.txt
# LiveKit Core
livekit-agents[openai,google,silero,deepgram]>=0.12.0
livekit>=0.18.0

# LLM & Embeddings
anthropic>=0.40.0
openai>=1.50.0
sentence-transformers>=3.0.0

# Vector Store
chromadb>=0.5.0

# Web Scraping
playwright>=1.45.0
beautifulsoup4>=4.12.0
requests>=2.32.0
lxml>=5.2.0
httpx>=0.27.0

# Text Processing
langchain-text-splitters>=0.3.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
pydantic>=2.9.0
aiohttp>=3.10.0
asyncio-throttle>=1.0.2
